Update
Created a system to store players in a party and then will have the player make a character if none are there. This wraps back around and will allow the player to choose multiple options. One being to make a roll. Allows for player to make a roll, add modifieres, chooss the dice type and the amount of dice to throw.