File io work
Continued working on outputting all data to a file, have also added a randomized height generator for the character in question. All of this gets printed to the terminal for immediate viewing and then it outputted into a file to then be used by anyone else. See text file "Ivan" for a finished product